Split out type definitions to their own header.
Avoids having to include vulkan.h in external replayer.
